cats/serializers.py
from rest_framework import serializers
from .models import SpyCat


class SpyCatSerializer(serializers.ModelSerializer):
    class Meta:
        model = SpyCat
        fields = ['id', 'name', 'years_of_expirience', 'breed', 'salary']
